Introduce "coroutines"
Not exactly coroutines, as only the most basic feature of a coroutine is
implemented. Also, there are some specific semantics to our application.

The point of this is to write a function with a delay very much like we
would in the straightforward case, but instead of blocking give control
back to the main loop.