Review USB, SD, setup and loop
This started tracking a bug in the interaction betwee USB and XBee. The
Xbee module is branched to the Socket0. Socket0 and USB share UART0.
The bug is, when USB log enabled and USB cable disconnected, the XBee
module thinks he's receiving a packet. Reproduced with v12, apparently
not with the v15.

From this bug I went to review as well the setup and boot process. It
should be more clear now.