Dont power sensors if not needed
Issues found:
(1) when using only BME-280 (I2C, 3V3) the 5V was powered on
(2) when using only CTD-10 (SDI-12, 5V) the 3V3 was powered on

This happens in TwoWire::secureBegin because WaspSensorAgr_v20.h was
included unconditionally.

This is related to bug https://github.com/Libelium/waspmoteapi/issues/23
and requires latest version of our waspmoteapi fork, which fixes it.

For best power saving the USE_XXX defines must be correctly set.