Remove old dead code, saves some program memory
This code was done to further reduce the battery usage. But battery
usage is very good now, and the code is quite complex (so a potential
source of instability) and it takes quite some memory.

This was actually deployed in the field, and seemed to work fine.

The code was deactivated a while ago, and now it's removed completely.