Set Xbee retries to 2
I reduced retries to 1 when introduced app level acks, to save battery
in bad weather conditions, since it's a waste to retry in that
situation.

But we're having issues with the frames not reaching the Pi from motes
that don't have a direct link. Setting retries to 2 may help.

Issue #54