4G reduce connection timeout to 30s (from 120s)
In the logs from Svalbard we have seen many reboots by the Watchdog
while the 4G module tries to connect. This is unnecessary as the same
logs show that connections usually take 5-6 seconds, so waiting for
2minutes is too much.

Still we're conservative and set the timeout to 30s. The Watchdog
reboots are not harmful either.