saveFrames: use system time, not RTC time
When I2C fails RTC time is zero, this leads to the 000000.TXT file. In
this case it's not better to use system time, since what we have to do
is to reboot and recover the I2C bus; but through the code anyway we
should use system time instead of RTC time.